Add capability to update what policy is applied to a container
This functionallity allows us to catch all new docker containers and apply a default policy to them, afterwords the user can choose which policy the container should follow.

Note: The implicit mounts currently won't carry over to the new policy. We can possibly track this according to container rather that policy so that they will follow accross the policy change.

Another note:
It would be great if policy could be added without having to restart the deamon.