Experimental - Load new policies every 10sec
This commit is incomplete and contains hardcoded values.

The idea here is that we want to be able to load new policies after the daemon is created. This is because Docker Containers create their own root filesystem, so if we want to enforce file system rules on a Docker Container, then we need to
1. Create the container
2. Lookup which overlay fs is in use (/var/lib/docker/overlay2/..../merged)
3. Use this as the root file path for filesytem rules for the policy
4. Load the new policy into the kernel
5. Apply the policy to the container